% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAssoc.R
\name{addGroups}
\alias{addGroups}
\alias{addWindows}
\title{Add group names or sliding windows to association test results}
\usage{
addGroups(iterator, x)

addWindows(iterator, x)
}
\arguments{
\item{iterator}{SeqVarIterator object}

\item{x}{assocTestSeq results}
}
\description{
Add group names or sliding windows to association test results
}
