% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhenotypes.R
\name{addInvNorm}
\alias{addInvNorm}
\title{Add inverse normal transform}
\usage{
addInvNorm(annot, nullmod, outcome, covars)
}
\arguments{
\item{annot}{Data frame with sample annotation}

\item{nullmod}{Null model from \pkg{\link[GENESIS]{GENESIS}}}

\item{outcome}{Name of outcome (for printing message)}

\item{covars}{Names of covariates (for printing message)}
}
\value{
\code{annot} with additional column "resid.norm"
}
\description{
Inverse normal transform the residuals of a null model, and add the results to annotation.
}
