% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAssoc.R
\name{addMAC}
\alias{addMAC}
\title{Add MAC column to association test output}
\usage{
addMAC(assoc, assoc_type)
}
\arguments{
\item{assoc}{results from \code{\link[GENESIS]{assocTestSingle}} or \code{\link[GENESIS]{assocTestAggregate}}}

\item{assoc_type}{Type of association test ("single", "aggregate", "window")}
}
\value{
\code{assoc} with a "MAC" column added to the results data.frame
}
\description{
Add MAC column to association test output
}
