% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantData.R
\name{addVariantData}
\alias{addVariantData}
\title{Add variantData to a SeqVarData object}
\usage{
addVariantData(gds, variants)
}
\arguments{
\item{gds}{A \code{\link{SeqVarData}} object}

\item{variants}{A data.frame with columns "variant.id", or ("chr", "pos", "ref", "alt")}
}
\value{
\code{gds} with data from \code{variants} in the \code{variantData} slot
}
\description{
Add variantData to a SeqVarData object
}
