% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAssoc.R
\name{addWindows}
\alias{addWindows}
\title{Add information on windows to a data.frame of results}
\usage{
addWindows(iterator, x)
}
\arguments{
\item{iterator}{SeqVarWindowIterator object}

\item{x}{data.frame of results}
}
\description{
Add information on windows to a data.frame of results
}
