% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAssoc.R
\name{assocFilterByFile}
\alias{assocFilterByFile}
\title{Filter association test results to just thsose specified by the user.}
\usage{
assocFilterByFile(assoc, varfile)
}
\arguments{
\item{assoc}{results from \code{\link{getAssoc}}}

\item{varfile}{RData file containing vector of \code{id} values from \code{assoc} to keep. See details.}
}
\value{
\code{assoc} with only the specified identifiers
}
\description{
Filter association test results to just thsose specified by the user.
}
\details{
If \code{varfile} contains a space, `chromosome` will be inserted and the variant ids included will be assumed to be from that chromosome.
}
