% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateLambda}
\alias{calculateLambda}
\title{Calculate lambda, the genomic inflation factor}
\usage{
calculateLambda(stat, df)
}
\arguments{
\item{stat}{Vector of test statistics}

\item{df}{Degrees of freedom}
}
\value{
lambda
}
\description{
Calculate lambda, the genomic inflation factor
}
