% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locuszoom.R
\name{calculateLD}
\alias{calculateLD}
\title{Calculate LD}
\usage{
calculateLD(gdsfile, variant.id, ref.var = NULL, sample.id = NULL)
}
\arguments{
\item{gdsfile}{character string with GDS file name}

\item{variant.id}{vector of variant.id}

\item{ref.var}{variant.id of reference variant (if NULL, LD is calculated pairwise for all variants)}

\item{sample.id}{vector of sample.id to use (if NULL, all samples)}
}
\value{
r^2
}
\description{
Calculate LD
}
