% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\docType{data}
\name{chromosomes_hg19}
\alias{chromosomes_hg19}
\alias{chromosomes_hg38}
\title{Start and end positions of each chromosome}
\format{
\code{\link[GenomicRanges]{GRanges}} object

An object of class \code{GRanges} of length 23.
}
\source{
UCSC Genome Browser
}
\usage{
chromosomes_hg19

chromosomes_hg38
}
\description{
Start and end positions of each chromosome
}
\examples{
\dontrun{
library(dplyr)
db <- src_mysql(dbname="hg19", host="genome-mysql.cse.ucsc.edu", user="genome")
chromInfo <- tbl(db, "chromInfo")

dat <- select(chromInfo, chrom, size) \%>\%
    collect() \%>\%
    filter(chrom \%in\% paste0("chr", c(1:22, "X"))) \%>\%
    mutate(chrom=factor(sub("chr", "", chrom), levels=c(1:22, "X"))) \%>\%
    arrange(chrom) \%>\%
    mutate(chrom=as.character(chrom))

library(GenomicRanges)
chromosomes_hg19 <- GRanges(seqnames=dat$chrom,
                            ranges=IRanges(start=1, end=dat$size))
}
}
\keyword{datasets}
