% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{constructFilename}
\alias{constructFilename}
\title{Construct filename from chromosome and segment}
\usage{
constructFilename(prefix, chromosome = NA, segment = NA)
}
\arguments{
\item{prefix}{Prefix of filename}

\item{chromosome}{chromosome}

\item{segment}{segment}
}
\value{
Character string of form "<prefix>_chr<chromosome>_seg<segment>.RData"
}
\description{
Construct filename from chromosome and segment
}
