% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countThreads.R
\name{countThreads}
\alias{countThreads}
\title{Count number of threads available to the R process}
\usage{
countThreads()
}
\value{
Number of threads
}
\description{
Count number of threads available to the R process
}
