% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{custom_render_markdown}
\alias{custom_render_markdown}
\title{Internal function to render markdown documents}
\usage{
custom_render_markdown(markdown_template_name, output_file, parameters = NULL)
}
\arguments{
\item{markdown_template_name}{The name of the markdown template, excluding ".Rmd"}

\item{output_file}{The name of the output file to create}

\item{parameters}{A named list of parameters to pass to the markdown document}
}
\description{
Internal function to render markdown documents
}
\details{
The markdown template is expected to live in the \code{inst/rmd} directory and an error
is raised if it is not found.
}
\seealso{
\code{\link[rmarkdown]{render}}
}
