% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{defineSegments}
\alias{defineSegments}
\title{Define segments for parallel processing of the genome}
\usage{
defineSegments(seg.length, n, build = c("hg19", "hg38"))
}
\arguments{
\item{seg.length}{Segment length in base pairs}

\item{n}{Number of segments to generate. Only used if \code{seg.length} is missing.
Minimum number of segments is 23 (one per chromosome).}

\item{build}{Genome build}
}
\value{
\code{\link[GenomicRanges]{GRanges}} object with segments covering the genome
}
\description{
Define segments for parallel processing of the genome
}
\seealso{
\code{\link{writeSegmentFile}}, \code{\link{subsetBySegment}}
}
