% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAssoc.R
\name{formatAssocSingle}
\alias{formatAssocSingle}
\title{Format single-variant assocation test results}
\usage{
formatAssocSingle(seqData, assoc)
}
\arguments{
\item{seqData}{A \code{\link[SeqArray]{SeqVarGDSClass}} object
(needed to get chromosome and position if not present)}

\item{assoc}{data.frame with assocation test results}
}
\value{
data.frame including standard columns ("variantID", "chr", "pos", "n", "MAF", "minor.allele")
}
\description{
Return association test results for single-variant tests in a standard format
}
\details{
Ensures that single-variant association test results from different sources
(\code{\link[GENESIS]{assocTestMM}}, \code{\link[SeqVarTools]{regression}})
will all have a standard format.
}
