% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gds2ibdobj}
\alias{gds2ibdobj}
\title{Read an ibdobj from a GDS file}
\usage{
gds2ibdobj(file, sample.id = NULL)
}
\arguments{
\item{file}{filename of GDS file}

\item{sample.id}{vector of sample.id}
}
\value{
Object of type "snpgdsIBDClass"
}
\description{
Read an ibdobj from a GDS file
}
