% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAssoc.R
\name{getAssoc}
\alias{getAssoc}
\title{Get association test results}
\usage{
getAssoc(files, assoc_type)
}
\arguments{
\item{files}{Vector of file names with association test results}

\item{assoc_type}{Type of association test ("single", "aggregate", "window")}
}
\value{
data.frame including standard columns ("chr", "pos", "start", "end", "stat", "pval", "MAC")
}
\description{
Return association test results in a standard format
}
\details{
Read association test results in multiple files and combine all into a single
data frame with standard column names.

If a single aggregate unit contains variants from multiple chromosomes, each chromosome will have its own row in the output.
}
