% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGRM.R
\name{getGRM}
\alias{getGRM}
\title{Load a GRM or kinship matrix}
\usage{
getGRM(config, sample.id)
}
\arguments{
\item{config}{Config object (named vector) with params "pcrelate_file", "grm_file"}

\item{sample.id}{Vector of samples to include}
}
\value{
List of GRMs or kinship matrices. \code{NULL} if both pcrelate_file and grm_file are \code{NA}.
}
\description{
Load a GRM or kinship matrix
}
