% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhenotypes.R
\name{getPhenotypes}
\alias{getPhenotypes}
\title{Get phenotypes}
\usage{
getPhenotypes(config)
}
\arguments{
\item{config}{Config object (named vector) with params "phenotype_file", "n_pcs", "pca_file",
"outcome", "covars", "sample_include_file", "gds_file", "conditional_variant_file"}
}
\value{
list with annot, outcome, covars, sample.id
}
\description{
Get phenotypes
}
