% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{getSegments}
\alias{getSegments}
\title{Read segments from a text file}
\usage{
getSegments(file)
}
\arguments{
\item{file}{File with column names "chromosome", "start", "end " in the header}
}
\value{
\code{\link[GenomicRanges]{GRanges}} object with segments
}
\description{
Read segments from a text file
}
\seealso{
\code{\link{defineSegments}}, \code{\link{writeSegmentFile}}, \code{\link{subsetBySegment}}
}
