% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insertChromString}
\alias{insertChromString}
\title{Format a string by inserting chromosome into a blank space}
\usage{
insertChromString(x, chr, err = NULL)
}
\arguments{
\item{x}{Character string}

\item{chr}{Chromosome number (or character) to instert}

\item{err}{If not \code{NULL}, print this string with an error message about requiring a blank space}
}
\value{
String \code{x} with \code{chr} inserted into blank space
}
\description{
Format a string by inserting chromosome into a blank space
}
