% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{intToChr}
\alias{intToChr}
\title{Integer chromosome code to character}
\usage{
intToChr(chr)
}
\arguments{
\item{chr}{Integer chromosome code}
}
\value{
Character chromosome (23 -> X, 24 -> Y)
}
\description{
Integer chromosome code to character
}
