% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kinobj}
\alias{kinobj}
\title{Return a kinship object for use in pcairPartition}
\usage{
kinobj(file)
}
\arguments{
\item{file}{filename with kinship object (.gds or .RData)}
}
\value{
If file is a GDS file, a GDS object providing a file connection. Otherwise, the R object stored in file.
}
\description{
Return a kinship object for use in pcairPartition
}
