% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list2gds}
\alias{list2gds}
\title{Write each element of a list to a GDS node}
\usage{
list2gds(x, file)
}
\arguments{
\item{x}{a list}

\item{file}{character string with filename of GDS file to create}
}
\description{
Write each element of a list to a GDS node
}
