% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhenotypes.R
\name{matchAnnotGds}
\alias{matchAnnotGds}
\title{Return an expanded AnnotatedDataFrame with the same sample.id as the gds_file}
\usage{
matchAnnotGds(gds, annot)
}
\arguments{
\item{gds}{SeqVarGDSClass object}

\item{annot}{AnnotatedDataFrame with sample.id matching gds file}
}
\value{
AnnotatedDataFrame with sample.id matching gds file, and NA values for all
  samples not in original annot
}
\description{
Return an expanded AnnotatedDataFrame with the same sample.id as the gds_file
}
