% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{medianKinship}
\alias{medianKinship}
\title{Return the median kinship value from a matrix}
\usage{
medianKinship(mat)
}
\arguments{
\item{mat}{A kinship matrix}
}
\value{
The median kinship value
}
\description{
Return the median kinship value from a matrix
}
