% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{modelString}
\alias{modelString}
\title{Create model strings}
\usage{
modelString(outcome, covars, random, group_var)
}
\arguments{
\item{outcome}{outcome string}

\item{covars}{vector of covariate strings}

\item{random}{vector of random effect strings}

\item{group_var}{grouping variable string}
}
\value{
String representing analysis model
}
\description{
Create model strings
}
