% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAssoc.R
\name{omitKnownHits}
\alias{omitKnownHits}
\title{Omit known hits from an association test data frame}
\usage{
omitKnownHits(assoc, hits, flank = 500)
}
\arguments{
\item{assoc}{data.frame with assocation test results (including columns chr, pos)}

\item{hits}{data.frame with known hits (including columns chr, pos)}

\item{flank}{Number of kb on either side of each known hit to exclude}
}
\value{
data.frame with assocation test results not in regions around known hits
}
\description{
Omit known hits from an association test data frame
}
