% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rankNorm}
\alias{rankNorm}
\title{Inverse normal transform}
\usage{
rankNorm(x)
}
\arguments{
\item{x}{Vector of values to transform}
}
\value{
Vector with transformed values of \code{x}
}
\description{
Inverse normal transform
}
