% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAssoc.R
\name{removeConditional}
\alias{removeConditional}
\title{Remove conditional variants from assoc file}
\usage{
removeConditional(assoc, varfile)
}
\arguments{
\item{assoc}{results from \code{\link[GENESIS]{assocTestSingle}}}

\item{varfile}{conditional variant filename (RData containing data.frame with columns "variant.id", "chr")}
}
\value{
\code{assoc} with conditional variants removed
}
\description{
Remove conditional variants from assoc file
}
