% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{sample_annotation}
\alias{sample_annotation}
\title{Example sample annotation}
\format{
\code{\link[Biobase]{AnnotatedDataFrame}} object
}
\usage{
sample_annotation
}
\description{
Example sample annotation
}
\keyword{datasets}
