% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{segments}
\alias{segments}
\title{Example segments}
\format{
\code{\link[GenomicRanges]{GRanges}} object
}
\usage{
segments
}
\description{
Example segments
}
\examples{
segments <- defineSegments(seg.length=1e7, build="hg19")
}
\keyword{datasets}
