% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullModel.R
\name{smallNullModel}
\alias{smallNullModel}
\title{Return a small version of the null model (no NxN matrices)}
\usage{
smallNullModel(nullmod)
}
\arguments{
\item{nullmod}{null model object}
}
\value{
null model object without large matrices
}
\description{
Return a small version of the null model (no NxN matrices)
}
