% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{subsetBySegment}
\alias{subsetBySegment}
\title{Subset list of variant groups by segment}
\usage{
subsetBySegment(varList, segment, segment.file)
}
\arguments{
\item{varList}{A list of data.frames, each with columns "chr" and "pos" OR a GRanges or GRangesList}

\item{segment}{An integer indicating which segment to select}

\item{segment.file}{The name of the file describing segments}
}
\value{
Subset of \code{varList} where the first variant is in the segment \code{segment}
}
\description{
Select groups of variants where the first variant in the group is in the requested segment
}
\seealso{
\code{\link{defineSegments}}, \code{\link{writeSegmentFile}}, \code{\link{filterBySegment}}, \code{\link{aggregateGRanges}}
}
