% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locuszoom.R
\name{writeBED}
\alias{writeBED}
\title{Write a bed file to use as a track in a LocusZoom plot}
\usage{
writeBED(x, file, track.label = "")
}
\arguments{
\item{x}{data.frame with columns "chr", "start", "end"}

\item{file}{output file name}
}
\description{
Write a bed file to use as a track in a LocusZoom plot
}
