% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locuszoom.R
\name{writeLD}
\alias{writeLD}
\title{Write a LD file to use in a LocusZoom plot}
\usage{
writeLD(x, ld, ref.var, file)
}
\arguments{
\item{x}{data.frame with columns "variant.id", "chr", "pos"}

\item{ld}{vector of LD}

\item{ref.var}{variant.id of the reference variant}

\item{file}{output file name}
}
\description{
Write a LD file to use in a LocusZoom plot
}
