% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locuszoom.R
\name{writeMETAL}
\alias{writeMETAL}
\title{Write a metal-format file with variants to use in a LocusZoom plot}
\usage{
writeMETAL(x, file)
}
\arguments{
\item{x}{data.frame with columns "chr", "pos", "pval"}

\item{file}{output file name}
}
\description{
Write a metal-format file with variants to use in a LocusZoom plot
}
