% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{writeSegmentFile}
\alias{writeSegmentFile}
\title{Write segments to a text file}
\usage{
writeSegmentFile(segments, file)
}
\arguments{
\item{segments}{\code{\link[GenomicRanges]{GRanges}} object with segments}

\item{file}{Output file name}
}
\description{
Write segments to a text file
}
\seealso{
\code{\link{defineSegments}}, \code{\link{getSegments}}, \code{\link{filterBySegment}}
}
