% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getobj}
\alias{getobj}
\title{Get an R object stored in an Rdata file}
\usage{
getobj(Rdata)
}
\arguments{
\item{Rdata}{path to an Rdata file containing a single R object to load}
}
\value{
The R object stored in \code{Rdata}
}
\description{
Returns an R object stored in an Rdata file
}
\details{
Loads an R object and stores it under a new name without creating a duplicate copy.
If multiple objects are stored in the same file, only the first one will be returned.
Supports files saved with \code{\link{saveRDS}}, provided the file extension ".rds" is used.
}
\examples{
x <- 1:10
file <- tempfile()
save(x, file=file)
y <- getobj(file)
unlink(file)

}
