% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{modelString}
\alias{modelString}
\alias{modelOutcomeString}
\title{Create model strings for an analysis}
\usage{
modelString(outcome, covars, random, group_var, inverse_normal = FALSE)

modelOutcomeString(outcome, inverse_normal = FALSE)
}
\arguments{
\item{outcome}{outcome string}

\item{covars}{vector of covariate strings}

\item{random}{vector of random effect strings}

\item{group_var}{grouping variable string}

\item{inverse_normal}{indicator of whether the model was refit with inverse-normal transformed residuals}
}
\value{
String representing analysis model
}
\description{
Create model strings for an analysis
}
\section{Functions}{
\itemize{
\item \code{modelString}: Create a string for the entire model

\item \code{modelOutcomeString}: Create model outcome strings
}}

