% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinPoints.R
\name{thinPoints}
\alias{thinPoints}
\title{Thin points in a data frame for plotting}
\usage{
thinPoints(dat, value, n = 10000, nbins = 10, groupBy = NULL)
}
\arguments{
\item{dat}{a data.frame}

\item{value}{the column in \code{dat} to sample on}

\item{n}{the number of points to select in each bin}

\item{nbins}{the number of bins to divide \code{dat$value} into}

\item{groupBy}{if not \code{NULL}, bins are calculated within each value of \code{groupBy} separately}
}
\value{
A subset of \code{dat}. The maxumum number of rows is \code{n*nbins*ngroup} where
  \code{ngroup} = the number of unique elements in \code{dat$groupBy}. The number of rows may be less
  if some bins have fewer than \code{n} elements.
}
\description{
Sample a data frame by sampling the value of one of the columns
}
\examples{
dat <- data.frame(x=1:100, y=sample(letters[1:2], 100, replace=TRUE))
thinPoints(dat, "x", n=2, nbins=5, groupBy="y")
}
\references{
\url{http://stackoverflow.com/questions/30950016/dplyr-sample-n-where-n-is-the-value-of-a-grouped-variable}
}
